requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
google-genai
python-jose[cryptography]
werkzeug
pillow
bcrypt
Jinja2
requests
httpx
pytest
pytest-asyncio
pytest-playwright
playwright
